@@ -1,6 +1,4 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>ch.alpine</groupId>
@@ -25,7 +23,7 @@
 		<dependency>
 			<groupId>ch.alpine</groupId>
 			<artifactId>java</artifactId>
-			<version>0.2.4</version>
+			<version>0.3.0</version>
 		</dependency>
 		<dependency>
 			<groupId>ch.alpine</groupId>
@@ -40,7 +38,7 @@
 		<dependency>
 			<groupId>com.formdev</groupId>
 			<artifactId>flatlaf</artifactId>
-			<version>1.6.4</version>
+			<version>2.0</version>
 		</dependency>
 
 		<dependency>
@@ -101,7 +99,7 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.8.1</version>
 				<configuration>
-					<release>11</release>
+					<release>17</release>
 				</configuration>
 			</plugin>
 			<plugin>
@@ -122,8 +120,7 @@
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-javadoc-plugin</artifactId>
 				<version>3.3.1</version>
-				<configuration>
-				</configuration>
+				<configuration></configuration>
 			</plugin>
 
 			<plugin>
