@@ -1,7 +1,7 @@
 // code by jph
 package ch.alpine.sophus.demo.lev;
 
-import java.util.Objects;
+import java.util.Optional;
 
 import ch.alpine.sophus.gds.ManifoldDisplay;
 import ch.alpine.sophus.hs.Biinvariant;
@@ -18,24 +18,14 @@ public enum Bitype {
   ;
 
   public Biinvariant from(ManifoldDisplay manifoldDisplay) {
-    switch (this) {
-    case METRIC1:
-    case METRIC2:
-      Biinvariant biinvariant = manifoldDisplay.metricBiinvariant();
-      return Objects.nonNull(biinvariant) //
-          ? biinvariant
-          : Biinvariants.LEVERAGES;
-    case LEVERAGES1:
-    case LEVERAGES2:
-      return Biinvariants.LEVERAGES;
-    case GARDEN:
-      return Biinvariants.GARDEN;
-    case HARBOR:
-      return Biinvariants.HARBOR;
-    case CUPOLA:
-      return Biinvariants.CUPOLA;
-    default:
-      throw new IllegalArgumentException();
-    }
+    return switch (this) {
+    case METRIC1, METRIC2 -> Optional.ofNullable(manifoldDisplay.metricBiinvariant()) //
+        .orElse(Biinvariants.LEVERAGES);
+    case LEVERAGES1, LEVERAGES2 -> Biinvariants.LEVERAGES;
+    case GARDEN -> Biinvariants.GARDEN;
+    case HARBOR -> Biinvariants.HARBOR;
+    case CUPOLA -> Biinvariants.CUPOLA;
+    default -> throw new IllegalArgumentException();
+    };
   }
 }
